progress on admin

+++ database/migrations/2024_07_13_183021_create_penduduk_sub_kriterias_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('penduduk_id');
             $table->unsignedBigInteger('sub_kriteria_id');
-            $table->float('nilai');
+            $table->double('nilai', 8, 2);
             $table->timestamps();
         });
     }
